requirements.txt
-i https://pypi.org/simple
argparse==1.4.0
boto3==1.20.14
botocore==1.23.14
click==8.0.3
dominate==2.6.0
environs==9.3.5
flask-bootstrap4==4.0.2
flask-sqlalchemy==2.5.1
flask-wtf==1.0.0
flask==2.0.2
greenlet==1.1.2 ; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
itsdangerous==2.0.1
jinja2==3.0.3
jmespath==0.10.0
markupsafe==2.0.1
marshmallow==3.14.1
psycopg2==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
sqlalchemy==1.4.27
urllib3==1.26.7
visitor==0.1.3
werkzeug==2.0.2
wtforms==3.0.0
